+++ server/removeObjectFromTrash_test.go
@@ -140,10 +140,15 @@ func TestUndeleteExpungedObjectFails(t *testing.T) {
 		Snippets: testhelpers.SnippetTP10,
 		Found:    true,
 	}
+	attributesResponse := aac.UserAttributesResponse{
+		Success:        true,
+		UserAttributes: "{\"diasUserGroups\":{\"projects\":[{\"projectName\":\"DCTC\",\"groupNames\":[\"ODrive\"]}]}}",
+	}
 
 	fakeAAC := &aac.FakeAAC{
-		SnippetResp: &snippetResponse,
-		Err:         nil,
+		SnippetResp:        &snippetResponse,
+		UserAttributesResp: &attributesResponse,
+		Err:                nil,
 	}
 
 	fakeDAO := &dao.FakeDAO{
@@ -196,10 +201,15 @@ func TestUndeleteObjectWithDeletedAncestorFails(t *testing.T) {
 		Snippets: testhelpers.SnippetTP10,
 		Found:    true,
 	}
+	attributesResponse := aac.UserAttributesResponse{
+		Success:        true,
+		UserAttributes: "{\"diasUserGroups\":{\"projects\":[{\"projectName\":\"DCTC\",\"groupNames\":[\"ODrive\"]}]}}",
+	}
 
 	fakeAAC := &aac.FakeAAC{
-		SnippetResp: &snippetResponse,
-		Err:         nil,
+		SnippetResp:        &snippetResponse,
+		UserAttributesResp: &attributesResponse,
+		Err:                nil,
 	}
 
 	fakeDAO := &dao.FakeDAO{